Move getPacketGroupSummaries from packet service to packet group service
