+++ api/app/src/main/kotlin/packit/service/PacketService.kt
@@ -13,11 +13,8 @@ import packit.contentTypes
 import packit.exceptions.PackitException
 import packit.helpers.PagingHelper
 import packit.model.*
-import packit.model.dto.PacketGroupSummary
-import packit.model.dto.toPacketGroupSummary
 import packit.repository.PacketGroupRepository
 import packit.repository.PacketRepository
-import packit.service.utils.getDisplayNameForPacket
 import java.security.MessageDigest
 import java.time.Instant
 
@@ -29,7 +26,6 @@ interface PacketService
     fun importPackets()
     fun getMetadataBy(id: String): PacketMetadata
     fun getFileByHash(hash: String, inline: Boolean, filename: String): Pair<ByteArrayResource, HttpHeaders>
-    fun getPacketGroupSummaries(pageablePayload: PageablePayload, filter: String): Page<PacketGroupSummary>
     fun getPacketsByName(
         name: String, payload: PageablePayload
     ): Page<Packet>
@@ -73,53 +69,6 @@ class BasePacketService(
     override fun getPackets(): List<Packet>
     {
         return packetRepository.findAll()
-    }
-
-    /**
-     * Find the id of the latest packet per group, excluding cases where the user does not have access.
-     *
-     * @param names The names of the packet groups.
-     * @return For each group, an id of the latest packet in the group.
-     */
-    private fun findLatestPacketIdsForGroups(names: List<String>): List<String>
-    {
-        return names.mapNotNull {
-            try {
-                packetGroupRepository.findLatestPacketIdForGroup(it)?.id
-            } catch (e: AccessDeniedException) {
-                null
-            }
-        }
-    }
-
-    /**
-     * This function, somewhat inefficiently, applies a filter to metadata from outpack_server, rather than in SQL.
-     * This is because we need to filter on the displayName, which is not stored in the database.
-     * We intend to move to a more efficient solution in the future once we have made architectural changes.
-     */
-    override fun getPacketGroupSummaries(
-        pageablePayload: PageablePayload,
-        filter: String
-    ): Page<PacketGroupSummary> {
-        val allPacketGroups = packetGroupRepository.findAll()
-        val packetIds = findLatestPacketIdsForGroups(allPacketGroups.map { it.name })
-        val allPacketsMetadata = outpackServerClient.getMetadata()
-        // Filter the metadata to include only the packets previously established as being the latest in their group,
-        // then, if the name or display name matches the search filter,
-        // calculate the packetCount and displayName to generate a PacketGroupSummary.
-        val latestPackets = allPacketsMetadata.filter { it.id in packetIds }
-            .mapNotNull {
-                val display = getDisplayNameForPacket(it.custom, it.name)
-
-                if (it.name.contains(filter, ignoreCase = true) || display.contains(filter, ignoreCase = true)) {
-                    val packetCount = allPacketsMetadata.count { p -> p.name == it.name }
-                    it.toPacketGroupSummary(packetCount, display)
-                } else {
-                    null
-                }
-            }.sortedByDescending { it.getLatestTime() }
-
-        return PagingHelper.convertListToPage(latestPackets, pageablePayload)
     }
 
     override fun getPacketsByName(name: String, payload: PageablePayload): Page<Packet>
